Tweaks to PIDF and DriveControls
I increased kI for the flywheel so it handles firing better (and slows down less when other mechanisms are running). I also changed TeleopShooter to take inputs from Shuffleboard (I'll rename the command later, since it's not going to be used for teleop in an actual match).

The driver controller should re-zero when DriveControls runs initialize(), so it doesn't get messed up during competition.