2-ish ball auto, shooting doesn't work though for some reason????
For reasons unknown, the robot seems to chronically undershoot, even though the same data points worked fire before. This happened before but we thought we fixed it, and maybe we did and this is something else but maybe it isn't and we never fixed it and we're going to have to redo all of the tables for like the third time. In conclusion, aaaaaaaaa?????