Switched to 02/28 tables and commented out unnecessary readouts
The CAN bus utilization was at 100%, which might or might not explain the bizarre issues we've been having (my guess is not, but we'll see). The old tables seem like they might work (which makes like negative amounts of sense but whatever), so I'm hoping these changes are at least kind of okay.