Changed status frame periods for high-CAN usage devices, worked on 2-ball auto
Apparently some CAN devices use up more for certain functionalities, so we lowered the status frame periods for the ones we don't really need (or can deal with being slower). It...kind of reduced the CAN bus utilization? A little?

Also, I worked on the two-ball auto, but it still undershoots and the feeding has some issues (it just doesn't stop the feeder fast enough for some reason, even when I raised the 'closed' threshold on the upper one to the point of almost being the same as the 'open' threshold). Maybe I need to redo some of that code, it's still very WIP.