Separated shooter and feeder, got working 2-ball auto
We had to move the feeder to its own subsystem, because the feeder had to run in parallel with the flywheel ramping up during autonomous. This required a lot of changes, most of which are minor but did mean we changed a lot of files. The 2-ball auto does work though, which is pretty great.