Range-based auto firing works!
Yeah!!!!! :D

(But seriously, it does work pretty well; the bug preventing it from shooting was (as always) that you can't set the hood every loop or the profile can't start, so we fixed that and it works as intended).

We tested that from a variety of ranges and it works quite well, as well as the exact points (and it performs a little better than the exact points because we're not relying on human input).