Tweaked shooting controls to avoid getting stuck
The only _logic_ change I made was requiring a brief period when the sensor needs to be closed for a bit in FeederLaunchCommand before the falling edge, to prevent one cargo from pushing another past the sensor and back and creating a falling edge that terminates the command. A lot of the sequences for auto-firing in teleop were adapted from the 2-ball auto, so I tried to change things to only automate the stuff that Evan can't do consistently on its own.