WEEK 3 COMMIT 2: Added in built-in hood offset for all targeting commands
Despite not changing the shooter code, the shooting is completely off from what it was previously, to the point where we could barely score. This occurs even for the non-vision-based targeting command, which strongly suggests that the mechanism itself is somehow behaving differently than it was before.

Our assessment of the shooter's behavior is that this is a systematic error, affecting all shooter configurations equally. In particular, the shots were much higher and steeper than they should have been, which suggests that the error is in the hood. The hood motor reaches its target position, so if this is the case the mechanism itself must be somehow behaving differently. Our hope is that if the error is systematic, offsetting all input hood angles by the same amount should counteract the issue regardless of what is causing it. I included a variable on the dashboard to control this offset, and as soon as we have time on the practice field tomorrow we intend on testing different values at different ranges, hopefully finding an offset that returns the shooter to its normal behavior.