WEEK 3 COMMIT 4: Final changes to shooting logic, climber controls, etc.
At least, final for the time being.

We anticipated needing to recalibrate, but it ended up just working. We still don't know why, though after we got back to the MPR we _did_ find out that our hub is actually 2 inches lower than it should be, which might explain a lot (certainly not everything, though).