Got 4-ball auto working significantly better, tweaked shooting
I made a bunch of changes, but the most notable ones are that the hood has an auto-indexing command that runs in autonomous and hopefully improves its ability to zero consistently, AbsoluteDriveCommand can drive backwards and has more parameters, and the flywheel current limit was entirely disabled.

Getting rid of current limits seems, like, a little questionable, but as soon as we did that the shooting got way more consistent (especially when firing two at once) so I wouldn't be surprised if that current limit was part of why the shooting was inconsistent. I also fixed some of the sequencing logic in the 4-ball auto, and it does seem to work better now (still an average of 3 but that's still pretty good compared to some teams at Week 3). Exciting!