FINAL UPDATE FROM AFTER WORLDS AAAAAAAAAAAAAA
Weird to think this is the last commit I'll actually do for competition code :'(

But seriously, this is the best available code that we ran at Houston for the last two or so days. We should at some point clean this up because I'd been doing quick fixes for things for the past two weeks and there's a lot of nonsense in there now, but we can do that later.

(Also I swear I am going to do hub localization for post-season if it kills me also it's not going to kill me because it's like stupidly easy if you have more than 15 consecutive minutes with the robot to work on it)